pyproject.toml
[tool.poetry]
name = "aiogramtemplate"
version = "0.1.0"
description = ""
authors = ["taimast <charlskenno@gmail.com>"]
packages = [{ include = "src" }]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
uvloop = { version = ">=0.17.0", markers = "sys_platform != 'win32'" }
aiogram = "^3.14.0"
pydantic = "^2.7.1"
loguru = "^0.7.0"
PyYAML = "^6.0"
APScheduler = { version = "^4.0.0a4", allow-prereleases = true }
sqlalchemy = "^2.0.9"
psycopg2 = "^2.9.6"
aiosqlite = "^0.19.0"
pydantic-settings = "^2.0.2"
alembic = "^1.13.0"
cachetools = "^5.3.2"
requests = "^2.31.0"
bs4 = "^0.0.2"
lxml = "^5.1.0"
multi-merchant = { git = "https://github.com/taimast/multi-merchant.git" }
sqladmin = { extras = ["full"], version = "^0.18.0" }
ftl-translator = { git = "https://github.com/taimast/ftl-translator.git" }
fluentogram = "^1.1.7"
aiohttp-socks = "^0.9.0"
redis = {extras = ["hiredis"], version = "^5.2.0"}
sqlmodel = "^0.0.22"
aiocsv = "^1.3.2"

[tool.poetry.group.dev.dependencies]
watchdog = ">=2.3.0,<3.0.0"
sqlalchemy-utils = "^0.41.1"
jurigged = "^0.5.7"


[tool.pyright]
venvPath = "."
venv = ".venv"
